http://stackoverflow.com/questions/2522299/c-catch-blocks-catch-exception-by-value-or-reference
The standard practice for exceptions in C++ is ...

Throw by value, catch by reference
Catching by value is problematic in the face of inheritance hierarchies. Suppose for your example that there is another type MyException which inherits from CustomException and overrides items like an error code. If a MyException type was thrown your catch block would cause it to be converted to a CustomException instance which would cause the error code to change.

References

http://www.oreillynet.com/pub/a/network/2003/05/05/cpluspocketref.html