pass fully formed json config to Python code. Deprecate sending separate 'main' and 'overrides' arguments.
This also fixes the issue that "Inherits" usage for PyPort config files didn't work.
It works now, because the entire inherited and overiden config is passed as a single string (the argument previously used as 'main' config.
And the 'override' argument is now passed as an empty string to avoid a breaking API change.

This only breaks Python code that relied on config being passed in separate arguments instead of just merging them anyway (like in the examples).